Keep stamp file on dry run

+++ src/stamp.rs
@@ -28,14 +28,18 @@ impl Timestamp {
     }
 
     /// Attempts to load the the timestamp file in the given directory.
-    pub fn load(target_dir: &Path) -> Result<Timestamp, Error> {
+    ///
+    /// On dry run, keep the timestamp file. Otherwise it is deleted.
+    pub fn load(target_dir: &Path, dry_run: bool) -> Result<Timestamp, Error> {
         let mut path = target_dir.to_path_buf();
         path.push("sweep.timestamp");
         let mut file =
             File::open(&path).context(format!("failed to read stamp file {}", path.display()))?;
         let mut contents = String::new();
         file.read_to_string(&mut contents)?;
-        remove_file(&path)?;
+        if !dry_run {
+            remove_file(&path)?;
+        }
         let timestamp: Timestamp = from_str(&contents)?;
         Ok(timestamp)
     }
